Updated like logic & better images grid

+++ app/src/main/java/com/sudzusama/vkimageclassifier/data/local/preferences/VKSessionPreferences.kt
@@ -11,11 +11,11 @@ class VKSessionPreferences @Inject constructor(private val preferences: SharedPr
     }
 
     fun getToken(): String? = preferences.getString(AUTH_TOKEN, "")
-    fun getUserId(): Long = preferences.getLong(USER_ID, -1)
+    fun getUserId(): Int = preferences.getInt(USER_ID, -1)
 
-    fun saveSession(token: String, userId: Long) {
+    fun saveSession(token: String, userId: Int) {
         preferences.edit { putString(AUTH_TOKEN, token) }
-        preferences.edit { putLong(USER_ID, userId) }
+        preferences.edit { putInt(USER_ID, userId) }
     }
 
     fun clearSession() {
